+++ assignment1_0.cpp
@@ -6,6 +6,8 @@
 #include "opencv2/highgui/highgui.hpp"
 #include "opencv2/imgproc/imgproc.hpp"
 #include "imgproc.h"
+#include "utils.h"
+#include "examples.h"
 
 int main() {
 	cv::Mat inputImage, outputImage;
