+++ imgproc.h
@@ -12,14 +12,4 @@ namespace IPCVL {
 		void calcHist_hs(cv::InputArray src, double histogram[][64]);
 		void backprojectHistogram(cv::InputArray src_hsv, cv::InputArray face_hsv, cv::OutputArray dst);
 	} // namespace IMG_PROC
-
-	namespace UTIL {
-		int quantize(int a);
-		double h_r(double model_hist[][64], double input_hist[][64], int j, int i);
-		void GetHistogramImage(int* histogram, cv::OutputArray dst, int hist_w, int hist_h);
-	} // namespace UTIL
-
-	namespace EXAMPLE {
-		void ChangeContrastAndBrightness(cv::InputArray src, cv::OutputArray dst, double alpha, int beta);
-	} // namespace EXAMPLE
 } // namespace IPCVL
